MCPJVM-408: complete Regression Suite runtime extraction from Artifact Spec
