+++ tools/core/tools-core/src/index.ts
@@ -5,3 +5,5 @@ export * from "./recipe_execution_plan.util";
 export * from "./recipe_intent_routing.util";
 export * from "./recipe_types.util";
 export * from "./synthesis_search_roots.util";
+export * from "./object_path_read";
+export * from "./placeholder_resolution";
