+++ tools/core/tools-core/src/object_path_read.ts
