+++ tools/core/tools-core/src/placeholder_resolution.ts
