+++ tools/features/execution-profile-export/loaders/export_manifest.loader.ts
@@ -2,7 +2,7 @@ import { promises as fs } from "node:fs";
 import path from "node:path";
 
 import type { ExecutionProfileExportManifest, ExecutionProfileExportPlanRun } from "../models/execution_profile_export.model";
-import { readProjectArtifact } from "@tools-project-artifact-spec/project_artifact.util";
+import { readProjectArtifact } from "@tools-feature-artifact-management";
 import { resolveRegressionPlansRootAbs } from "../../../spec/regression-execution-plan-spec/src/regression_artifact_paths.util";
 import { readExecutionOrchestrationSuiteResult } from "@tools-feature-regression-suite";
