+++ tools/spec/regression-execution-plan-spec/package.json
@@ -3,6 +3,6 @@
   "private": true,
   "version": "0.1.0",
   "description": "Artifact Spec models, validation, and path policy for execution-plan Artifacts",
-  "main": "src/regression_artifact_paths.util.ts",
-  "types": "src/regression_artifact_paths.util.ts"
+  "main": "src/index.ts",
+  "types": "src/index.ts"
 }
