+++ tools/spec/regression-execution-plan-spec/src/index.ts
@@ -1,0 +1,12 @@
+export * from "./models/regression_execution_plan_spec.model";
+export * from "./models/regression_execution_profile_export.model";
+export * from "./models/regression_replay_invocation.model";
+export * from "./models/regression_run_artifact.model";
+export * from "./models/regression_runtime_suite.model";
+export * from "./models/regression_transport.model";
+export {
+  validateExternalVerificationContract,
+  validateNormalizedExternalVerificationResultShape,
+} from "./external_verification_contract.util";
+export { validateCanonicalPlanContextKeys } from "./suite_context_key_validation.util";
+export { resolveRegressionPlansRootAbs } from "./regression_artifact_paths.util";
